Add unit tests for negation normal form, formula construction
Added unit tests for negation normal form for FOF formulas.
Added unit tests for formula creation to isolate a bug found
when testing nnf.
There appears to be a bug in the creation of quantifiers when a
bound variable does not appear or is bound twice (see the tests
for details).