For efficiency, separate typechecking into its own transformer
Typechecking & sanitizing used to be done when constructing a theory.
This meant it was being done at every single transformation, significantly
increasing memory usage (because of sanitizing) and causing frequent
invocations of the garbage collector.
Now, typechecking & sanitizing are only done at the beginning of the
compliation process and have been placed in a transformer.