Fix sort inference bug
When computing the substitution from the general theory to the original theory,
it would fail if a sort never was used in a constant or function but only
appeared on annotated variables of quantifiers.
Now, the substitution computation recurses over corresponding term pairs
to ensure it gets these sorts as well.