Parse CVC5 'as' as a Var
So it turns out that the "SMT value" concept as referenced throughout
SMTLIBCliSolver and SmtModelVisitor is surprisingly generic: DEs are
already mapped to "SMT values" through smtValue2DomainElement, which are
represented as strings of the getText() of the function definition
bodies for the DEs. The only restriction was that elsewhere, the SMT
values are assumed to be represented by Vars named that getText(). So
generate such a Var for the (as x y) construction as a hack to get this
working.